+++ app/scripts/api/snapshot/Transporter.js
@@ -52,7 +52,8 @@ define(function(require) {
         var toBeApplied = options.toBeApplied || {};
 
         //The list of change listeners
-        var changeListeners = [];
+        var changeListeners = {};
+        var configChangeListeners = {};
         var initialSetupCompleteListeners = [];
         var handshakeListeners = [];
 
@@ -159,11 +160,32 @@ define(function(require) {
             }
         };
 
+        function removeChangeListener(id) {
+            delete changeListeners[id];
+        }
+
         this.addChangeListener = function(changeListener) {
-            changeListeners.push(changeListener);
+            var id = uuid();
+            changeListeners[id] = changeListener;
+            return removeChangeListener.bind(this, id);
         };
+
         this.removeAllChangeListeners = function() {
-            changeListeners = [];
+            changeListeners = {};
+        };
+
+        function removeConfigChangeListener(id) {
+            delete configChangeListeners[id];
+        }
+
+        this.addConfigChangeListener = function(changeListener) {
+            var id = uuid();
+            configChangeListeners[id] = changeListener;
+            return removeConfigChangeListener.bind(this, id);
+        };
+
+        this.removeAllConfigChangeListeners = function() {
+            configChangeListeners = {};
         };
 
         /*
@@ -421,6 +443,7 @@ define(function(require) {
         var handleConfigChangeMessage = function(message) {
             if (message.handshake.authToken === handshake.authToken) {
                 handshake.config = message.handshake.config;
+                callConfigChangeListeners(handshake.config);
             }
         };
 
@@ -488,6 +511,9 @@ define(function(require) {
                     }
                     pendingMessages.forHandshake = [];
                 }
+
+
+                callConfigChangeListeners(handshake.config);
             }
         };
 
@@ -739,6 +765,13 @@ define(function(require) {
         var callChangeListeners = function(values) {
             _.each(changeListeners, function(changeListener) {
                 changeListener(values);
+            });
+        };
+
+        // Calls all the configChangeListeners
+        var callConfigChangeListeners = function(config) {
+            _.each(configChangeListeners, function(changeListener) {
+                changeListener(config);
             });
         };